# Slightly better error package performance.