--- a/registry/client/blob_writer_test.go
+++ b/registry/client/blob_writer_test.go
@@ -94,7 +94,7 @@ func TestUploadReadFrom(t *testing.T) {
 								"detail": "more detail"
 							}
 						]
-					} `),
+					}`),
 			},
 		},
 		// Test 400 invalid json