@stevvooe 2015-10-16
Delaying formatting may cause issues with thread safety. The danger is passing that error to another goroutine and having an object be evaluated that is no longer valid or in scope.

To tell you the truth, I'd like to completely remove `WithArgs` from the error interface. It's a silly method and isn't very safe. These worked much faster when they were literally just error codes.

Was this causing a specific performance issue?