@calavera 2015-10-16
> Was this causing a specific performance issue?

we've seen methods in docker which performance has decreased by ~50ms only because they initialize errors.

> To tell you the truth, I'd like to completely remove WithArgs from the error interface. It's a silly method and isn't very safe. These worked much faster when they were literally just error codes.

I'm all for removing code.