@stevvooe 2015-10-19
> we've seen methods in docker which performance has decreased by ~50ms only because they initialize errors.

Could you cite real cases in this analysis? Something sounds very wrong if a simple allocation of a small struct is causing a 50ms performance problem.